+++ test/fixtures/command/yarn.yml
@@ -1,5 +1,6 @@
 expected_dependency: autoprefixer
-source_path: test/fixtures/yarn
-cache_path: test/fixtures/yarn/.licenses
 sources:
   yarn: true
+
+apps:
+  - source_path: test/fixtures/yarn/v1
